requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1
django-import-export==4.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
setuptools==72.2.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
uritemplate==4.1.1

